数字/requirements.txt
torch
torchvision
numpy
matplotlib
gradio
Pillow
